requirement.txt
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.64.0
PyYAML>=6.0
Pillow>=9.0.0
faiss-cpu>=1.7.4